modularized, fixed some nitpicks
