+++ apps/nexus-web/src/features/community-showcase/components/DesktopShowcase.tsx
@@ -1,0 +1,92 @@
+/* eslint-disable @next/next/no-img-element */
+
+import { Container, Stack, Text } from "@packages/spark-ui";
+import { FeaturedEventCard } from "./FeaturedEventCard";
+import { PastEventsCarousel } from "./PastEventsCarousel";
+
+/**
+ * DesktopShowcase
+ *
+ * Full desktop layout (rendered only at the `md` breakpoint and above).
+ * Includes:
+ *   - Background decoration (blobs, Cirby, stardust, horizon)
+ *   - Section heading
+ *   - FeaturedEventCard (with modal trigger)
+ *   - PastEventsCarousel
+ */
+
+interface DesktopShowcaseProps {
+  onOpenModal: () => void;
+}
+
+export function DesktopShowcase({ onOpenModal }: DesktopShowcaseProps) {
+  return (
+    <>
+      {/* ── Background decorations (desktop only) ── */}
+      <div className="hidden md:block">
+        <div className="w-[802px] h-[811px] right-0 top-0 absolute bg-pink-400/30 rounded-full blur-[400px] translate-x-1/2 -translate-y-1/2" />
+        <div className="w-[802px] h-[811px] left-0 top-0 absolute bg-blue-500/25 rounded-full blur-[400px] -translate-x-1/2 translate-y-1/2" />
+        <div className="w-[966px] h-[977px] left-0 top-0 absolute bg-sky-400/20 rounded-full blur-[400px] -translate-x-1/5 translate-y-3/4" />
+        <div className="w-[966px] h-[977px] left-0 bottom-0 absolute bg-pink-400/30 rounded-full blur-[400px] translate-x-1/2 translate-y-4/20" />
+        <img
+          className="w-[30vw] max-w-[493px] h-auto right-0 top-55 absolute -mr-[6vw]"
+          src="/community-showcase/community-showcase-cirby.webp"
+          alt=""
+        />
+        <img
+          className="w-[1354.15px] h-[1656.86px] left-0 top-0 absolute opacity-75 -ml-165 mt-40"
+          src="/community-showcase/community-showcase-space-dust-1.webp"
+          alt=""
+        />
+        <img
+          className="w-[1400px] h-auto right-20 bottom-235 absolute opacity-75 translate-x-1/3"
+          src="/community-showcase/community-showcase-space-dust-2.webp"
+          alt=""
+        />
+      </div>
+
+      {/* ── Main content ── */}
+      <Container>
+        <Stack gap="2xl" className="hidden md:flex relative z-10 gap-32">
+          {/* Section heading */}
+          <Stack gap="md" className="items-center">
+            <Text
+              variant="heading-1"
+              gradient="white-blue"
+              align="center"
+              weight="bold"
+              className="z-20"
+            >
+              Community Showcase
+            </Text>
+            <Text
+              as="h2"
+              variant="heading-4"
+              weight="bold"
+              align="center"
+              className="text-white z-10"
+            >
+              Discover what our community has been building together.
+            </Text>
+          </Stack>
+
+          {/* Horizon illustration */}
+          <img
+            className="w-[min(1700px,140vw)] h-auto max-w-none left-1/2 -translate-x-1/2 -top-30 absolute"
+            src="/community-showcase/community-showcase-horizon.webp"
+            alt=""
+          />
+
+          {/* Featured event card — 1st instance */}
+          <FeaturedEventCard onOpenModal={onOpenModal} />
+
+          {/* Featured event card — 2nd instance (intentional duplicate) */}
+          <FeaturedEventCard onOpenModal={onOpenModal} />
+
+          {/* Past events carousel */}
+          <PastEventsCarousel />
+        </Stack>
+      </Container>
+    </>
+  );
+}
