+++ apps/nexus-web/src/features/community-showcase/index.ts
@@ -1,7 +1,10 @@
 /**
  * Community Showcase Feature Module
  *
- * Exports all components needed to render the community showcase page.
+ * Public API: exports the top-level section component and shared types.
+ * Internal sub-components (MobileShowcase, DesktopShowcase, etc.) are
+ * implementation details and intentionally not re-exported here.
  */
 
 export { CommunityShowcaseSection } from "./components/CommunityShowcaseSection";
+export type { PastEvent } from "./types";
