+++ apps/nexus-web/src/features/community-showcase/types.ts
@@ -1,0 +1,9 @@
+/**
+ * Community Showcase Feature — Shared Types
+ */
+
+export interface PastEvent {
+  date: string;
+  title: string;
+  image: string;
+}
